gitlab modules: minor refactor (#6384)
* gitlab modules: minor refactor

* add changelog frag

* Update plugins/module_utils/gitlab.py


* Update plugins/module_utils/gitlab.py


* update changelog frag

* remove extraneous bracket

---------
