+++ changelogs/fragments/6384-gitlab-refactor.yml
@@ -1,0 +1,2 @@
+minor_changes:
+  - gitlab_group_variable, gitlab_project_variable - refactor function out to module utils (https://github.com/ansible-collections/community.general/pull/6384).
